Better retries with concurrent requests
Previously, if two or more threads completed with 401/403 errors concurrently before
the `updateAccessToken` happened, only one of them would be retried.

Now, they'll all be retried. That multiple requests have to be retried due to the same
expired token is an unavoidable side-effect of the fact that we allow multiple
concurrent requests (and that we only know that the token is expired as a result from
the request itself), and we can live with that tradeoff.

[PLAT-5219](https://vertexvis.atlassian.net/browse/PLAT-5219)